# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Mono-Api"
    },
    "host": "localhost:8081",
    "basePath": "/techmillresource/mono-api/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "uikey": {
            "in": "header",
            "name": "uikey",
            "type": "apiKey"
        },
        "session_key": {
            "in": "header",
            "name": "apisessionkey",
            "type": "apiKey"
        }
    },
    "paths": {
        "/getAudit": {
            "post": {
                "tags": [
                    "audit"
                ],
                "summary": "retrieve audit by userId in system",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "list of audit functions by userId in system",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/GetAudit"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetAudit"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
        "/getUserAudit": {
            "post": {
                "tags": [
                    "audit"
                ],
                "summary": "retrieve audit for user in system",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "list of user's audit functions in system",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/GetUserAudit"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetUserAudit"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
...
        "/getInfolog": {
            "post": {
                "tags": [
                    "logs"
                ],
                "description": "retrieve info data logged in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "list of info data logged by date",
                    "schema": {
                        "$ref": "#/definitions/GetInfolog"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetInfolog"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
        "/getErrorlog": {
            "post": {
                "tags": [
                    "logs"
                ],
                "description": "retrieve error(s) logged in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "list of error(s) logged by date",
                    "schema": {
                        "$ref": "#/definitions/GetErrorlog"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetErrorlog"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
        "/getDebuglog": {
            "post": {
                "tags": [
                    "logs"
                ],
                "description": "retrieve debug data logged in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "list of debug data logged by date",
                    "schema": {
                        "$ref": "#/definitions/GetDebuglog"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetDebuglog"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
...
        "/getBank": {
            "post": {
                "tags": [
                    "bank"
                ],
                "description": "user(s) bank details in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "bank details of user in system",
                    "schema": {
                        "$ref": "#/definitions/GetBank"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetBank"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
        "/createBank": {
            "post": {
                "tags": [
                    "bank"
                ],
                "description": "user bank details to create in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "bank details of user to create in system",
                    "schema": {
                        "$ref": "#/definitions/CreateBank"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/CreateBank"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
        "/updateBank": {
            "post": {
                "tags": [
                    "bank"
                ],
                "description": "user bank details to update in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "bank details of user to update in system",
                    "schema": {
                        "$ref": "#/definitions/UpdateBank"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UpdateBank"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
...
    },
    "definitions": {
        "GetAudit": {
            "type": "object",
            "required": [
                "profileId",
                "userId",
                "startDate",
                "endDate",
                "channel",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "userId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "GetUserAudit": {
            "type": "object",
            "required": [
                "profileId",
                "startDate",
                "endDate",
                "channel",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
...
        "GetInfolog": {
            "type": "object",
            "required": [
                "startDate",
                "endDate"
            ],
            "properties": {
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                }
            }
        },
        "GetErrorlog": {
            "type": "object",
            "required": [
                "startDate",
                "endDate"
            ],
            "properties": {
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                }
            }
        },
        "GetDebuglog": {
            "type": "object",
            "required": [
                "startDate",
                "endDate"
            ],
            "properties": {
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                }
            }
        },
...
        "GetBank": {
            "type": "object",
            "required": [
                "profileId",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "CreateBank": {
            "type": "object",
            "required": [
                "profileId",
                "bankName",
                "accountNumber",
                "accountName",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "bankName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "accountNumber": {
                    "type": "number",
                    "uniqueItems": true
                },
                "accountName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "UpdateBank": {
            "type": "object",
            "required": [
                "profileId",
                "bankId",
                "bankName",
                "accountNumber",
                "accountName",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "bankId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "bankName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "accountNumber": {
                    "type": "number",
                    "uniqueItems": true
                },
                "accountName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
...
    }
}